% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.R
\name{anova.reg}
\alias{anova.reg}
\alias{case.names.reg}
\alias{coef.reg}
\alias{deviance.reg}
\alias{df.residual.reg}
\alias{formula.reg}
\alias{model.frame.reg}
\alias{nobs.reg}
\alias{plot.reg}
\alias{predict.reg}
\alias{print.reg}
\alias{residuals.reg}
\alias{simulate.reg}
\alias{summary.reg}
\alias{terms.reg}
\alias{variable.names.reg}
\alias{vcov.reg}
\alias{weights.reg}
\title{Methods for objects of class \dQuote{reg}}
\usage{
\method{anova}{reg}(object, ...)

\method{case.names}{reg}(object, ...)

\method{coef}{reg}(object, ...)

\method{deviance}{reg}(object, ...)

\method{df.residual}{reg}(object, ...)

\method{formula}{reg}(x, ...)

\method{model.frame}{reg}(formula, ...)

\method{nobs}{reg}(object, ...)

\method{plot}{reg}(x, ...)

\method{predict}{reg}(object, ...)

\method{print}{reg}(x, ...)

\method{residuals}{reg}(object, ...)

\method{simulate}{reg}(object, ...)

\method{summary}{reg}(object, ...)

\method{terms}{reg}(x, ...)

\method{variable.names}{reg}(object, ...)

\method{vcov}{reg}(object, ...)

\method{weights}{reg}(object, ...)
}
\arguments{
\item{object}{the same}

\item{\dots}{Additional arguments passed to methods. In the case of \code{print}, this is passed through to \code{\link[stats]{printCoefmat}}.}

\item{x}{An object returned by \code{\link{reg}}}

\item{formula}{the same}

\item{model}{the same}
}
\value{
An object returned by the underlying method
}
\description{
Methods for common modelling generics
}
\details{
These functions simply provide a convenience wrapper allowing for code of the form \code{predict(mod)} rather than \code{predict(mod$model)}. These functions per se do nothing directly.
}
