% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.R
\name{reg}
\alias{reg}
\title{Stata-style Regression Modelling Function}
\usage{
reg(data, formula, ..., vcov_type = if (is.null(vcov_cluster)) "const" else
  "HC0", vcov_cluster = NULL, boot_iterations = 1000L, digits = 2L,
  signif.stars = FALSE)
}
\arguments{
\item{data}{A data frame, or \dQuote{survey design} object from the survey package.}

\item{formula}{A model formula.}

\item{\dots}{Additional arguments passed to \code{\link[stats]{glm}} or \code{\link[survey]{svyglm}}.}

\item{vcov_type}{A character string specifying a variance-covariance estimation procedure. See \code{\link[sandwich]{vcovHC}}.}

\item{vcov_cluster}{If non-\code{NULL}, a formula specifying a clustering variable to be used for clustered standard errors.}

\item{boot_iterations}{If \code{vcov_type = "boot"}, an integer specifying the number of bootstrap iterations.}

\item{digits}{An integer specifying the preferred number of digits to use when printing the output (passed to \code{\link{print.model}}).}

\item{signif.stars}{A logical specifying whether to print significance stars. Passed to \code{\link{print.model}}.}
}
\value{
A \dQuote{reg} object, which contains a traditional R modelling object in element \dQuote{model}.
}
\description{
This is a slim wrapper around various regression modelling functions that simplifies the process of estimating regression equations with various robust variance-covariance estimation procedures. This negates the need for the two-step exercise of \code{summary(lm(...))} to view results. It also supports survey design objects through the same interface. Reversing the typical formula-data argument order to data-argument order makes the function easily usable within a data analysis \dQuote{pipeline}.
}
\details{
Estimation is provided by \code{\link[stats]{glm}} and \code{\link[survey]{svyglm}}, respectively.
}
\seealso{
\code{\link{summary.reg}}
}
